+++ adapters/lumi/__init__.py
@@ -2,6 +2,7 @@ from adapters.on_off_switch_adapter import OnOffSwitchAdapter
 from adapters.dimmable_ct_bulb_adapter import DimmableCtBulbAdapter
 from adapters.generic.motion_sensor import MotionSensorAdapter
 from adapters.generic.smoke_sensor import SmokeSensorAdapter
+from adapters.generic.temp_hum_sensor import TemperatureHumiditySensorAdapter
 from adapters.generic.water_leak_sensor import WaterLeakSensorAdapter
 from adapters.lumi.sensor_cube import SensorCube
 from adapters.lumi.sensor_magnet import SensorMagnet
@@ -11,7 +12,6 @@ from adapters.lumi.plug import Plug
 from adapters.lumi.QBKG03LM import QBKG03LM
 from adapters.lumi.QBKG12LM import QBKG12LM
 from adapters.lumi.JTQJBF01LM import JTQJBF01LM
-from adapters.lumi.WSDCGQ01LM import WSDCGQ01LM
 from adapters.lumi.WSDCGQ11LM import WSDCGQ11LM
 from adapters.lumi.WXCJKG12LM import WXCJKG12LM
 from adapters.lumi.WXCJKG13LM import WXCJKG13LM
@@ -38,7 +38,7 @@ lumi_adapters = {
     'RTCGQ01LM': MotionSensorAdapter,   # Xiaomi MiJia human body movement sensor
     'RTCGQ11LM': SensorMotionAq2,       # Xiaomi Aqara human body movement and illuminance sensor
     'SJCGQ11LM': WaterLeakSensorAdapter,# Xiaomi Aqara water leak sensor
-    'WSDCGQ01LM': WSDCGQ01LM,           # Xiaomi MiJia temperature & humidity sensor
+    'WSDCGQ01LM': TemperatureHumiditySensorAdapter, # Xiaomi MiJia temperature & humidity sensor
     'WSDCGQ11LM': WSDCGQ11LM,           # Xiaomi Aqara temperature, humidity and pressure sensor
     'WXCJKG12LM': WXCJKG12LM,           # Xiaomi Aqara Opple 4-button switch
     'WXCJKG13LM': WXCJKG13LM,           # Xiaomi Aqara Opple 6-button switch